Make EcoMug and BXDecay0 private dependencies
Remage's event generators do not provide a public interface that
exposes their internals such as EcoMug and BXDecay0. This makes
these package implementation details which a consumer of libremage
does not need to be aware of.

Hide EcoMug and BxDecay0 use behind compiler firewalls using a Pimpl.
Link BXDecay0 privately to libremage.