Minimal .clang-tidy with useful, safe checks as example
Tiny set of checks to demo functionality and that are generally safe
to apply automatically.

If we configure the project in a directory that's not a sibling of
the source directory, clang-tidy may fail to pick up the right
.clang-tidy file. As we will generally always run clang-tidy by
pointing it to our build directory, copy the .clang-tidy
file there by default so that file is picked up.